+++ PXCTestKit/CommandLineInterface.swift
@@ -27,7 +27,7 @@ import FBSimulatorControl
                        Option<Locale>("locale", Locale(identifier: "en"), description: "Locale to set for the Simulator."),
                        Option<Preferences>("preferences", Preferences(), description: "Path to some preferences.json to be applied with the Simulator."),
                        Option<Reporter>("reporter", .rspec, description: "Console reporter type. Supported values: rspec, json"),
-                       Option<Only>("only", Only(), description: "Comma separated list of tests that should be executed only. Example: TestClass1/testFoo,TestClass2/testBar"),
+                       VaradicOption<Only>("only", [], description: "Comma separated list of tests that should be executed only. Format: TARGET[:Class/case[,Class2/case2]]"),
                        VaradicOption<Destination>("destination", [], description: "A comma-separated set of key=value pairs describing the destination to use, just like xcodebuild -destination."),
                        Option<Double>("timeout", 3600.0, description: "Timeout in seconds for the test execution to finish."),
                        Flag("no-color", description: "")
@@ -43,7 +43,11 @@ import FBSimulatorControl
                     environment: ProcessInfo.processInfo.environment,
                     preferences: preferences.dictionary,
                     reporterType: reporter.type,
-                    testsToRun: only.testsToRun,
+                    testsToRun: only.reduce([String: Set<String>](), {
+                        var result = $0
+                        result[$1.targetName] = $1.testsToRun
+                        return result
+                    }),
                     simulators: destination.map({ $0.simulatorConfiguration }),
                     timeout: timeout,
                     consoleOutput: consoleOutput,