+++ PXCTestKitTests/Fixtures/Sample.source/SuccessfulTests/Tests.m
@@ -1,0 +1,29 @@
+//
+//  Tests.m
+//  SuccessfulTests
+//
+//  Created by Plunien, Johannes(AWF) on 30/11/16.
+//  Copyright © 2016 Johannes Plunien. All rights reserved.
+//
+
+#import <XCTest/XCTest.h>
+
+@interface SuccessfulTests : XCTestCase
+
+@end
+
+@implementation SuccessfulTests
+
+- (void)testOne {
+    XCTAssertTrue(YES);
+}
+
+- (void)testTwo {
+    XCTAssertTrue(YES);
+}
+
+- (void)testThree {
+    XCTAssertTrue(YES);
+}
+
+@end