+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Support abort via SIGINT
 * Support JSON stream reporter (similar to xctool's JSON stream)
+* Breaking change: --only expects a new format: TARGET[:Class/case[,Class2/case2]]
 
 # 0.2.6